moved to projects with userId

+++ src/main/java/ru/karelin/tmweb/service/ProjectService.java
@@ -1,5 +1,6 @@
 package ru.karelin.tmweb.service;
 
+import org.jetbrains.annotations.Nullable;
 import ru.karelin.tmweb.entity.Project;
 import ru.karelin.tmweb.repository.ProjectRepository;
 
@@ -7,9 +8,9 @@ import java.util.List;
 
 public class ProjectService {
     private final ProjectRepository projectRepository = ProjectRepository.getInstance();
-    private static final ProjectService ourInnstance = new ProjectService();
+    private static final ProjectService ourInstance = new ProjectService();
     public static ProjectService getInstance(){
-        return ourInnstance;
+        return ourInstance;
     }
     private ProjectService(){}
 
@@ -24,8 +25,17 @@ public class ProjectService {
     public void save(Project project) {
         projectRepository.save(project);
     }
-    public void remove(String projectId){
-        Project project = projectRepository.find(projectId);
+    public void remove(@Nullable String projectId, String userId){
+        if(projectId==null || projectId.isEmpty()) return;
+        Project project = projectRepository.findByIdAndUserId(projectId, userId);
         if (project!=null) projectRepository.remove(project);
+    }
+
+    public List<Project> findAllByUserId(String id) {
+       return projectRepository.findAllByUserId(id);
+    }
+
+    public Project findByIdAndUserId(String id, String userId) {
+        return projectRepository.findByIdAndUserId(id, userId);
     }
 }
